refactor: extract MtpDraftRunner to modeling layer, decouple from LLMInferWrapper
